Changed shipment destination to string
